Astra-50-"Particle Renderer Access Violation"

See merge request jszajek/astra!27

+++ Astra.Core/src/math/MathsUtils.h
@@ -41,7 +41,7 @@ namespace Astra::Math
 
 	static inline float Random()
 	{
-		return static_cast <float> (rand()) / static_cast <float> (RAND_MAX);
+		return static_cast <float> (rand()) / RAND_MAX;
 	}
 
 	static inline int RandomRange(int min, int max)
